feat(streams): generate stream types from the OpenAPI contract
Brings the streams service under codegen, the way documents already is.
`types.gen.ts` now holds the wire types, and `types.ts` keeps only what
codegen cannot express.

`codegen.json` needs two things beyond the usual service config:

- `ignore` for `streams/{streamId}/records`, since the records endpoints live
  under the streams path and would otherwise be pulled in here
- `relevantReferenceNames` roots that exclude `StreamResponseItem`, which
  carries `createdTime`. The API returns that as an epoch timestamp, but
  `getDateProps()` converts it to a `Date`, so the generated type would
  describe the wrong thing. Keeping it out of `types.gen.ts` leaves the name
  free for the hand-written `Stream`. Same approach as annotations.

`StreamWrite`, `StreamCreateSettings` and `StreamSettings` become aliases over
the generated types, and `StreamLifecycleSettings`, `StreamLimitSettings` and
`StreamLimit` are now generated outright - all shape-identical to what they
replace.

No public type is removed.
