+++ packages/stable/src/api/streams/types.gen.ts
@@ -1,0 +1,126 @@
+// Do not modify this file!
+// It was generated by the command "yarn codegen".
+// Instead update the code generation logic or the OpenAPI document.
+
+/**
+ * @example {"items":[{"externalId":"my-stream","settings":{"template":{"name":"BasicArchive"}}}]}
+ */
+export interface CreateStreamRequest {
+  /**
+   * streamItems
+   * List of streams.
+   *
+   */
+  items: StreamRequestItem[];
+}
+/**
+ * Stream settings which should be applied to a stream.
+ */
+export interface CreateStreamSettings {
+  /**
+   * Reference to a template which should be used to define initial settings for the stream.
+   *
+   */
+  template: {
+    name: string;
+  };
+}
+/**
+* Data lifecycle settings. These settings are populated from the stream creation template
+and there is no easy way to change them. This information is meant to be human-readable
+and is not really meant for machine consumption.
+*/
+export interface StreamLifecycleSettings {
+  /**
+   * dataDeletedAfter
+   * This setting is available only for immutable streams.
+   * ISO-8601 formatted time specifying how long to retain a record in this stream.
+   * After this time passes, records are scheduled to be removed from the stream.
+   * Record removal happens in batches, and some time may pass between a record
+   * being scheduled for deletion and it actually being deleted. During this time,
+   * the record will remain available, and contributing to
+   * consumption of limits.
+   *
+   * @example P1Y
+   */
+  dataDeletedAfter?: string;
+  /**
+   * retainedAfterSoftDelete
+   * Time until the soft deleted stream will actually be deleted by the system,
+   * in an ISO-8601 compliant date-time format. Data cannot be ingested to,
+   * or queried from, a soft-deleted stream. If a stream has been deleted by mistake,
+   * it can still be recovered from soft-deleted state.
+   *
+   * @example P42D
+   */
+  retainedAfterSoftDelete: string;
+}
+export interface StreamLimit {
+  /**
+   * consumed
+   * Amount of resource consumed.
+   *
+   * This value will be returned only in a Retrieve stream response,
+   * and only if ```includeStatistics``` request parameter is ```true```.
+   * Currently, this value is available for ```maxRecordsTotal``` and
+   * ```maxGigaBytesTotal``` limits only.
+   * @example 999
+   */
+  consumed?: number;
+  /**
+   * provisioned
+   * Amount of resource provisioned.
+   *
+   * @example 1000
+   */
+  provisioned: number;
+}
+/**
+ * Limits and usage.
+ */
+export interface StreamLimitSettings {
+  /**
+   * maxFilteringInterval
+   * Maximum length of time that the ```lastUpdatedTime``` filter can retrieve records for,
+   * in ISO-8601 format. This setting is only available for immutable streams.
+   *
+   * @example P1Y
+   */
+  maxFilteringInterval?: string;
+  /**
+   * maxGigaBytesTotal
+   * Maximum amount of data that can be stored in the stream, in gigabytes.
+   *
+   */
+  maxGigaBytesTotal: StreamLimit;
+  /**
+   * maxRecordsTotal
+   * Maximum number of records that can be stored in the stream.
+   *
+   */
+  maxRecordsTotal: StreamLimit;
+}
+/**
+ * Stream object.
+ */
+export interface StreamRequestItem {
+  /**
+   * streamId
+   * Stream identifier. The identifier must be unique within the project and must be a valid stream identifier.
+   * Stream identifiers can only consist of alphanumeric characters, hyphens, and underscores.
+   * It must not start with cdf_ or cognite_, as those are reserved for Cognite internal use.
+   * Stream id cannot be logs or records. Max length is 100 characters.
+   *
+   * @pattern ^[a-z]([a-z0-9_-]{0,98}[a-z0-9])?$
+   * @example test_stream_1
+   */
+  externalId: string;
+  settings: CreateStreamSettings;
+}
+/**
+ * Stream settings.
+ */
+export interface StreamResponseItemSettings {
+  lifecycle: StreamLifecycleSettings;
+  limits: StreamLimitSettings;
+}
