+++ packages/stable/src/api/streams/types.ts
@@ -1,38 +1,34 @@
 // Copyright 2025 Cognite AS
 
 import type { CreatedTime } from '../../types';
+import type {
+  CreateStreamSettings,
+  StreamRequestItem,
+  StreamResponseItemSettings,
+} from './types.gen';
 
 /**
  * Stream write - the input type for creating a stream
  */
-export interface StreamWrite {
-  /**
-   * Stream identifier. Must be unique within the project.
-   */
-  externalId: string;
-  /**
-   * Stream settings which should be applied to the stream
-   */
-  settings: StreamCreateSettings;
-}
+export type StreamWrite = StreamRequestItem;
 
 /**
  * Settings for creating a stream
  */
-export interface StreamCreateSettings {
-  /**
-   * Reference to a template which should be used to define initial settings for the stream
-   */
-  template: {
-    /**
-     * The stream settings template name
-     */
-    name: string;
-  };
-}
+export type StreamCreateSettings = CreateStreamSettings;
 
 /**
- * A stream - target for high volume data ingestion
+ * Stream settings returned from the API
+ */
+export type StreamSettings = StreamResponseItemSettings;
+
+/**
+ * A stream - target for high volume data ingestion.
+ *
+ * Not generated: the API returns `createdTime` as an epoch timestamp, which the
+ * SDK converts to a `Date` before handing it back. See the `relevantReferenceNames`
+ * roots in `codegen.json`, which keep `StreamResponseItem` out of `types.gen.ts`
+ * so this name stays free.
  */
 export interface Stream extends CreatedTime {
   /**
@@ -51,66 +47,6 @@ export interface Stream extends CreatedTime {
    * Stream settings
    */
   settings: StreamSettings;
-}
-
-/**
- * Stream settings returned from the API
- */
-export interface StreamSettings {
-  /**
-   * Data lifecycle settings
-   */
-  lifecycle: StreamLifecycleSettings;
-  /**
-   * Limits and usage
-   */
-  limits: StreamLimitSettings;
-}
-
-/**
- * Data lifecycle settings
- */
-export interface StreamLifecycleSettings {
-  /**
-   * Time specifying how long to retain a record in this stream.
-   */
-  dataDeletedAfter?: string;
-  /**
-   * Time until the soft deleted stream will actually be deleted by the system,
-   */
-  retainedAfterSoftDelete: string;
-}
-
-/**
- * Limits and usage for a stream
- */
-export interface StreamLimitSettings {
-  /**
-   * Maximum number of records that can be stored in the stream
-   */
-  maxRecordsTotal: StreamLimit;
-  /**
-   * Maximum amount of data that can be stored in the stream, in gigabytes
-   */
-  maxGigaBytesTotal: StreamLimit;
-  /**
-   * Maximum length of time that the lastUpdatedTime filter can retrieve records for
-   */
-  maxFilteringInterval?: string;
-}
-
-/**
- * Stream limit with provisioned and consumed values
- */
-export interface StreamLimit {
-  /**
-   * Amount of resource provisioned
-   */
-  provisioned: number;
-  /**
-   * Amount of resource consumed
-   */
-  consumed?: number;
 }
 
 /**
